+++ packages/web/src/javascripts/Components/Preferences/Panes/Account/SubscriptionSharing/Invite/InviteSuccess.tsx
@@ -1,0 +1,11 @@
+import { FunctionComponent } from 'react'
+
+const InviteSuccess: FunctionComponent = () => {
+  return (
+    <div>
+      <div className={'mb-2 font-bold text-info'}>Your invitation has been successfully sent.</div>
+    </div>
+  )
+}
+
+export default InviteSuccess